+++ Tmain/list-roles.d/run.sh
@@ -11,17 +11,22 @@ title()
     echo '#'
 }
 
+ignore_xml()
+{
+    grep -v Glade
+}
+
 title ''
-${CTAGS} --quiet --options=NONE --_list-roles=
+${CTAGS} --quiet --options=NONE --_list-roles= | ignore_xml
 
 title 'all:*'
-${CTAGS} --quiet --options=NONE --_list-roles='all:*'
+${CTAGS} --quiet --options=NONE --_list-roles='all:*' | ignore_xml
 
 title 'C:*'
 ${CTAGS} --quiet --options=NONE --_list-roles='C:*'
 
 title 'all:h'
-${CTAGS} --quiet --options=NONE --_list-roles='all:h'
+${CTAGS} --quiet --options=NONE --_list-roles='all:h' | ignore_xml
 
 title 'Sh:s'
 ${CTAGS} --quiet --options=NONE --_list-roles='Sh:s'