+++ docs/news.rst
@@ -27,6 +27,7 @@ The following parsers have been added:
 * Diff
 * DTS
 * Falcon
+* Glade *libxml*
 * Go
 * JSON
 * ObjectiveC
@@ -43,6 +44,7 @@ The following parsers have been added:
 
 See "Option library" about  *optlib*.
 See "External parser command" about *xcmd*.
+Libxml2 is needed to use the parser(s) marked with *libxml*.
 
 TIPS: you can list newly introduced parsers if you have
 exuberant-ctags with following command line: